Add automatic deps generation for Qt libraries in Linux
Implementation of a new feature that allows the qt_local_repo
rule to analyze the `.so` files from Qt libraries to determine
which other Qt libraries they depend on, and add them to
the `deps` parameter of the generated `cc_library` targets.

This feature is controlled by a new attribute `infer_deps`,
which defaults to False.

This has two caveats:
1. For the moment it is implemented only for Linux, using `ldd`
2. It does not distinguish between optional and non-optional
dependencies.

For example, in Qt6.5 the Quick3D module has an optional
dependency on ShaderTools. It still works if ShaderTools is not
installed, but because ShaderTools is linked to it (and there is
no way to know optional dependencies before runtime) it is added
as a dependency. This means users who use this feature will
need to install all optional dependencies for the modules they
have already installed.

Signed-off-by: Diego Ortín <dieortin@gmail.com>